+++ Twitter.php
@@ -5,7 +5,7 @@ namespace TijsVerkoyen\Twitter;
  * Twitter class
  *
  * @author		Tijs Verkoyen <php-twitter@verkoyen.eu>
- * @version		2.3.0
+ * @version		2.3.1
  * @copyright	Copyright (c), Tijs Verkoyen. All rights reserved.
  * @license		BSD License
  */
@@ -23,7 +23,7 @@ class Twitter
     const SECURE_API_PORT = 443;
 
     // current version
-    const VERSION = '2.3.0';
+    const VERSION = '2.3.1';
 
     /**
      * A cURL instance